Move the fizz code into its own package
