+++ quic/CMakeLists.txt
@@ -72,6 +72,7 @@ add_subdirectory(client)
 add_subdirectory(codec)
 add_subdirectory(common)
 add_subdirectory(congestion_control)
+add_subdirectory(fizz)
 add_subdirectory(flowcontrol)
 add_subdirectory(handshake)
 add_subdirectory(happyeyeballs)
