+++ quic/client/handshake/FizzClientExtensions.h
@@ -10,7 +10,7 @@
 
 #include <fizz/client/ClientExtensions.h>
 #include <quic/client/handshake/ClientTransportParametersExtension.h>
-#include <quic/handshake/FizzTransportParameters.h>
+#include <quic/fizz/handshake/FizzTransportParameters.h>
 
 namespace quic {
