+++ quic/client/handshake/FizzClientHandshake.cpp
@@ -10,7 +10,7 @@
 
 #include <quic/client/handshake/FizzClientExtensions.h>
 #include <quic/client/handshake/FizzClientQuicHandshakeContext.h>
-#include <quic/handshake/FizzBridge.h>
+#include <quic/fizz/handshake/FizzBridge.h>
 
 #include <fizz/protocol/Protocol.h>
