+++ quic/fizz/CMakeLists.txt
@@ -1,0 +1,6 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+add_subdirectory(handshake)
