+++ quic/fizz/handshake/FizzCryptoFactory.h
@@ -8,8 +8,8 @@
 
 #pragma once
 
+#include <quic/fizz/handshake/QuicFizzFactory.h>
 #include <quic/handshake/CryptoFactory.h>
-#include <quic/handshake/QuicFizzFactory.h>
 
 namespace quic {
