+++ quic/server/handshake/test/ServerHandshakeTest.cpp
@@ -26,9 +26,9 @@
 #include <quic/QuicConstants.h>
 #include <quic/client/handshake/FizzClientExtensions.h>
 #include <quic/common/test/TestUtils.h>
-#include <quic/handshake/FizzBridge.h>
+#include <quic/fizz/handshake/FizzBridge.h>
+#include <quic/fizz/handshake/QuicFizzFactory.h>
 #include <quic/handshake/HandshakeLayer.h>
-#include <quic/handshake/QuicFizzFactory.h>
 #include <quic/server/handshake/AppToken.h>
 #include <quic/server/handshake/ServerHandshake.h>
 #include <quic/state/StateData.h>
